Items now have proper creation and mod time set
API was always had separated creation and modification
times on Items, but pre-Java7 did not expose those,
hence as workaround they were always carrying the
same "lastModified" timestamp. This is not quite
practical in case of directories, that have this
timestamp changed whenever a change occurs within
it.

Now, the StorageItem#getCreated and StorageItem#getModified
are carrying the actual values corresponding to those
events.